+++ src/pulumi/core-dtap/storage/datalake.py
@@ -2,14 +2,15 @@ from pulumi import ResourceOptions, Output
 import pulumi_azure_native as azure_native
 import pulumi_azure as azure_classic
 
-from ingenii_azure_data_platform.utils import generate_resource_name
 from ingenii_azure_data_platform.defaults import STORAGE_ACCOUNT_DEFAULT_FIREWALL
-from ingenii_azure_data_platform.iam import (
-    GroupRoleAssignment,
-    UserAssignedIdentityRoleAssignment,
-)
 
-from project_config import azure_client, platform_config, platform_outputs
+from ingenii_azure_data_platform.iam import GroupRoleAssignment, UserAssignedIdentityRoleAssignment
+from ingenii_azure_data_platform.logs import log_diagnostic_settings, \
+    log_network_interfaces
+from ingenii_azure_data_platform.utils import generate_resource_name
+
+from logs import log_analytics_workspace
+from project_config import platform_config, platform_outputs
 from platform_shared import (
     shared_services_provider,
     get_devops_principal_id,
@@ -87,6 +88,37 @@ outputs["id"] = datalake.id
 outputs["name"] = datalake.name
 
 # ----------------------------------------------------------------------------------------------------------------------
+# DATA LAKE -> LOGGING
+# ----------------------------------------------------------------------------------------------------------------------
+
+log_diagnostic_settings(
+    platform_config, log_analytics_workspace.id, datalake.type,
+    datalake.id, datalake_name,
+    logs_config=datalake_config.get("logs", {}),
+    metrics_config=datalake_config.get("metrics", {})
+)
+
+blob_logs_and_metrics = datalake_config.get("storage_type_logging", {}) \
+                                       .get("blob", {})
+log_diagnostic_settings(
+    platform_config, log_analytics_workspace.id, datalake.type, 
+    datalake.id.apply(lambda dl_id: f"{dl_id}/blobservices/default"),
+    f"{datalake_name}-blob",
+    logs_config=blob_logs_and_metrics.get("logs", {}),
+    metrics_config=blob_logs_and_metrics.get("metrics", {})
+)
+
+table_logs_and_metrics = datalake_config.get("storage_type_logging", {}) \
+                                        .get("table", {})
+log_diagnostic_settings(
+    platform_config, log_analytics_workspace.id, datalake.type,
+    datalake.id.apply(lambda dl_id: f"{dl_id}/tableservices/default"),
+    f"{datalake_name}-table",
+    logs_config=table_logs_and_metrics.get("logs", {}),
+    metrics_config=table_logs_and_metrics.get("metrics", {})
+)
+
+# ----------------------------------------------------------------------------------------------------------------------
 # DATA LAKE -> PRIVATE ENDPOINTS
 # ----------------------------------------------------------------------------------------------------------------------
 
@@ -111,6 +143,18 @@ blob_private_endpoint = azure_native.network.PrivateEndpoint(
     custom_dns_configs=[],
     resource_group_name=resource_groups["infra"].name,
     subnet=azure_native.network.SubnetArgs(id=vnet.privatelink_subnet.id),
+)
+
+# To Log Analytics Workspace
+blob_private_endpoint_details = \
+    datalake_config.get("network", {}) \
+                   .get("private_endpoint", {}) \
+                   .get("blob", {})
+log_network_interfaces(
+    platform_config, log_analytics_workspace.id,
+    blob_private_endpoint_name, blob_private_endpoint.network_interfaces,
+    logs_config=blob_private_endpoint_details.get("logs", {}),
+    metrics_config=blob_private_endpoint_details.get("metrics", {})
 )
 
 # BLOB PRIVATE DNS ZONE GROUP
@@ -149,6 +193,18 @@ dfs_private_endpoint = azure_native.network.PrivateEndpoint(
     resource_group_name=resource_groups["infra"].name,
     custom_dns_configs=[],
     subnet=azure_native.network.SubnetArgs(id=vnet.privatelink_subnet.id),
+)
+
+# To Log Analytics Workspace
+dfs_private_endpoint_details = \
+    datalake_config.get("network", {}) \
+                   .get("private_endpoint", {}) \
+                   .get("dfs", {})
+log_network_interfaces(
+    platform_config, log_analytics_workspace.id,
+    dfs_private_endpoint_name, dfs_private_endpoint.network_interfaces,
+    logs_config=dfs_private_endpoint_details.get("logs", {}),
+    metrics_config=dfs_private_endpoint_details.get("metrics", {})
 )
 
 # DFS PRIVATE DNS ZONE GROUP